Always rebuild venvs when rebuilding wheels
The var 'repo_build_venv_force_rebuild' was made available to
allow the venv rebuilding to be optional, and defaulted to
false. This causes a mismatch between expected behaviour and the
actual behaviour.

When executing a repo-build, the deployer expectes the wheels
and venvs to be rebuilt. The actual behaviour is that the wheels
are rebuilt, but the venvs are left alone.

If a deployer/developer is updating SHA's for anything within the
same tag (or following the OSA master branch) then the services
never get deployed with updated code as the venvs never change.

This patch removes this var altogether. Deployers may still use
tag skipping to isolate tasks they wish to run if that's
required.

Change-Id: I8ef8d8c13e067223eb908fd2177cef8beb929397