Make the use of the repo index overrides optional
Currently the repo build process hard codes the use of a set of indexes
when building wheels. This overrides any existing pip configuration in
the repo container and is somewhat unintuitive as a result.

As the default index is pypi, which is the default for pip anyway, this
is a redundant hard-code which really shouldn't be required.

The specification of additional indexes should also be optional, as the
default of using pypi again just results in a duplicated check by pip.

Also, for context, OpenStack-CI gating will prefer the use of a
system-wide pip.conf which will be copied from the host to the container.
Without this patch, that simple solution to our gating process will not
work.

Change-Id: I25e4faecbffad6301509c62f55c34ba86b92a658