Remove global-requirements build from the build process
Currently the build process builds all the wheels for all packages
listed in OpenStack's global requirements.

This is a wasteful process as it contains packages needed by all
OpenStack projects, instead of just the projects OpenStack-Ansible
deploys in the integrated build.

This patch removes the wheel build of all global requirements and
switches the repo build process to only building the wheels that
are actually needed.

Depends-On: I6536993523f02ba49da741b9ea21535eb33fe402
Change-Id: I5a865391803de2beef49c61b55277fad6317561f