Add the galera_client dependency to the repo build
This change is required because without it the client libraries
used to build the MySQL-Python packages will depend on the base
system and they will be wrong.

This is an example of the issue:
  http://paste.openstack.org/show/492788/

This will cause multi node production issues when building a
standalong repository outside of the the OpenStack gate.

Closes-Bug: #1565128
Change-Id: I6536993523f02ba49da741b9ea21535eb33fe402
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>
(cherry picked from commit 9e6b78a9b459eebbae87c97ea474c724f50bd3e3)